+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 This package was formerly known as `fmutool`.
 
 # Version 1.9.3.1
-* ADDED: Highlight FMUs nested in selected container
-* ADDED: Tooltips for the parameters of containers
+* ADDED: GUI: contextual menu of FMUs let the user to launch fmueditor, fmutool or FMPy (if installed)
+* ADDED: GUI: Highlight FMUs nested in selected container
+* ADDED: GUI: Tooltips for the parameters of containers
+* ADDED: GUI: Add link is easier when FMU contains lot of variable. The parameters are displayed in italic.
+* ADDED: GUI: Import is able to load SSP file
 * FIXED: FMU extraction of a container
 
 # Version 1.9.3